#3364 move the only test that runs proxy from "statics" to "proxy" module
Starting a proxy slows down "statics" module's tests.
